refactor: フォント関連ファイルを src/font/ ディレクトリに集約
src/ ルート直下に散在していたフォント関連の8ファイル（+ テスト6ファイル）を
src/font/ に移動し、バレルエクスポート (index.ts) を新設。
モジュール境界を明確化しディレクトリの見通しを改善。
