+++ src/converter.ts
@@ -4,13 +4,13 @@ import { DEFAULT_OUTPUT_WIDTH } from "./utils/constants.js";
 import type { SlideElement } from "./model/shape.js";
 import type { LogLevel } from "./warning-logger.js";
 import { initWarningLogger, flushWarnings } from "./warning-logger.js";
-import { setTextMeasurer, resetTextMeasurer } from "./text-measurer.js";
+import { setTextMeasurer, resetTextMeasurer } from "./font/text-measurer.js";
 import { parsePptxData, parseSlideWithLayout } from "./pptx-data-parser.js";
-import type { FontMapping } from "./font-mapping.js";
-import { createFontMapping } from "./font-mapping.js";
-import { setFontMapping, resetFontMapping } from "./font-mapping-context.js";
-import { createOpentypeSetupFromSystem } from "./opentype-helpers.js";
-import { setTextPathFontResolver, resetTextPathFontResolver } from "./text-path-context.js";
+import type { FontMapping } from "./font/font-mapping.js";
+import { createFontMapping } from "./font/font-mapping.js";
+import { setFontMapping, resetFontMapping } from "./font/font-mapping-context.js";
+import { createOpentypeSetupFromSystem } from "./font/opentype-helpers.js";
+import { setTextPathFontResolver, resetTextPathFontResolver } from "./font/text-path-context.js";
 
 export interface ConvertOptions {
   /** 変換対象のスライド番号 (1始まり)。未指定で全スライド */
