+++ src/font/opentype-text-measurer.ts
@@ -1,5 +1,5 @@
 import type { TextMeasurer } from "./text-measurer.js";
-import { measureTextWidth as defaultMeasureTextWidth } from "./utils/text-measure.js";
+import { measureTextWidth as defaultMeasureTextWidth } from "../utils/text-measure.js";
 
 const PX_PER_PT = 96 / 72;
 const BOLD_FACTOR = 1.05;
